[PRP-337] Moving checkfile out into its own file, fixing consistent name of default import from logging.server
