+++ participant/app/utils/s3.server.ts
@@ -16,21 +16,13 @@ import {
 } from "app/utils/config.server";
 import { PassThrough } from "stream";
 import { writeAsyncIterableToWritable } from "@remix-run/node";
-import { fileTypeFromBuffer } from "file-type";
-import type { FileCheckResult } from "app/types";
-import {
-  MAX_UPLOAD_SIZE_BYTES,
-  S3_PRESIGNED_URL_EXPIRATION,
-} from "./config.server";
+
+import { S3_PRESIGNED_URL_EXPIRATION } from "./config.server";
 import { trimStart } from "lodash";
 import { File } from "@remix-run/node/dist/fetch";
-import logging from "app/utils/logging.server";
+import logger from "app/utils/logging.server";
 
 const PATHSTYLE = ENDPOINT_URL ? true : false;
-
-const validSize = (size: number): boolean => {
-  return size < MAX_UPLOAD_SIZE_BYTES;
-};
 
 export const parseKeyFromS3URL = (s3URL: string) => {
   const parsedURL = new URL(s3URL);
@@ -56,7 +48,7 @@ export const getFileFromS3 = async (key: string): Promise<File | undefined> => {
     if (error instanceof NotFound) {
       return undefined;
     }
-    logging.error(
+    logger.error(
       { location: "s3.server", type: "error.getFile", key: key, error: error },
       `Unable to get ${key} from S3: ${error}`
     );
@@ -82,7 +74,7 @@ export const readFileHeadFromS3 = async (
     if (error instanceof NotFound) {
       return undefined;
     }
-    logging.error(
+    logger.error(
       {
         location: "s3.server",
         type: "error.readFileHead",
@@ -110,7 +102,7 @@ export const headFilesizeFromS3 = async (
     if (error instanceof NotFound) {
       return undefined;
     }
-    logging.error(
+    logger.error(
       {
         location: "s3.server",
         type: "error.headFilesize",
@@ -143,7 +135,7 @@ export const getURLFromS3 = async (
     if (error instanceof NotFound) {
       return undefined;
     }
-    logging.error(
+    logger.error(
       { location: "s3.server", type: "error.getURL", key: key, error: error },
       `Unable to get URL for ${key}: ${error}`
     );
@@ -158,7 +150,7 @@ export const deleteFileFromS3 = async (key: string) => {
   });
   try {
     await s3Connection.send(command);
-    logging.info(
+    logger.info(
       { location: "s3.server", type: "info.deleteFile", key: key },
       `🗑️  Deleted ${key}`
     );
@@ -167,7 +159,7 @@ export const deleteFileFromS3 = async (key: string) => {
     if (error! instanceof NotFound) {
       return;
     } else {
-      logging.error(
+      logger.error(
         {
           location: "s3.server",
           type: "error.deleteFile",
@@ -179,49 +171,6 @@ export const deleteFileFromS3 = async (key: string) => {
       throw new Error(`Unable to delete ${key}: ${error}`);
     }
   }
-};
-
-// @TODO move checkFile() into a separate file.
-// Importing s3.server.ts outside of remix (such as in prisma seed scripts)
-// is blocked because the `file-type` package doesn't want to import properly.
-export const checkFile = async (key: string): Promise<FileCheckResult> => {
-  const fileSize = await headFilesizeFromS3(key);
-  if (!fileSize) {
-    return { error: "notFound" };
-  } else if (!validSize(fileSize)) {
-    return { error: "invalidSize", size: fileSize };
-  }
-
-  const startOfFile = await readFileHeadFromS3(key);
-  if (!startOfFile) {
-    logging.warn(
-      {
-        location: "s3.server",
-        type: "checkfile.cannotRead",
-        key: key,
-        fileSize: fileSize,
-      },
-      `❌ Unable to get head of file for ${key}`
-    );
-    return { error: "cannotRead", size: fileSize };
-  }
-  const fileType = await fileTypeFromBuffer(startOfFile);
-  if (!fileType) {
-    logging.warn(
-      {
-        location: "s3.server",
-        type: "checkfile.cannotType",
-        key: key,
-        fileSize: fileSize,
-      },
-      `❌ Unable to determine filetype for ${key}`
-    );
-    return { error: "cannotType", size: fileSize };
-  }
-  if (fileType.mime.includes("image") || fileType.mime == "application/pdf") {
-    return { mimeType: fileType.mime, size: fileSize };
-  }
-  return { mimeType: fileType.mime, error: "invalidType", size: fileSize };
 };
 
 // Thank you 🙏🏻 to https://github.com/remix-run/examples/issues/163
@@ -257,7 +206,7 @@ export async function uploadStreamToS3(data: any, filename: string) {
         return file.Location;
       }
     } catch (e) {
-      logging.error(
+      logger.error(
         { location: "s3.server", type: "uploadStream", filename: filename },
         `⚠️ File upload failed: ${e}; retrying ${
           retries + 1
